Integrator/SQL work from evening meeting of Aug 3
